% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bisquare_fns}
\alias{bisquare_1d}
\alias{bisquare_2d}
\alias{bisquare_B}
\alias{bisquare_fns}
\title{Bisquare function in one dimension}
\usage{
bisquare_1d(h, delta = 0, r = 1, A = 1)

bisquare_2d(h1, h2, delta = c(0, 0), r = 1, A = 1)

bisquare_B(h1, h2, delta = c(0, 0), r = 1, A = 1, areas = rep(1,
  length(h1)), n1 = 10L, n2 = 10L)
}
\arguments{
\item{h}{displacement (1d)}

\item{delta}{shift parameter(s)}

\item{r}{aperture parameter}

\item{A}{gain parameter (amplitude)}

\item{h1}{first component of displacement vector (2d)}

\item{h2}{second component of displacement vector (2d)}

\item{areas}{area associated with each column in B}

\item{n1}{number of rows}

\item{n2}{number of columns}
}
\description{
The bisquare function
}
\details{
The bisquare function is given by \deqn{b(s,v) \equiv \left\{\begin{array}{ll} A\{1 - (|v- s - \Delta|/r)^2\}^2, &| v -s  - \Delta| \le r \\ 0, & \textrm{otherwise}. \end{array} \right.}{b(s,v) =  A{1 - (|v- s - d|/r)^2}^2  if | v -s  - d| <= r,  0 otherwise}
The function \code{bisquare_1d} accepts \code{h} in any shape or size, while for \code{bisquare_2d}, \code{h1} and \code{h2} need to be vectors of the same size. The parameter \code{delta} needs to be equal to one or two in length, depending on the spatial dimension.

The function \code{bisquare_B} is used to construct the matrix \eqn{B} given the bisquare parameters. It is meant to be used in problems of 2 dimensions.
}
\examples{
h <- seq(-10,10,0.1)
y <- bisquare_1d(h=h,delta=3,r=4)
plot(h,y)
}

