% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initFEbasis}
\alias{initFEbasis}
\title{Initialise a finite element basis}
\usage{
initFEbasis(p, t, K)
}
\arguments{
\item{p}{\code{n} \eqn{\times} 2 matrix of vertex locations.}

\item{t}{\code{m} \eqn{\times} 3 matrix of triangulations. Each row identifies which rows of \code{p} make up each triangle.}

\item{K}{\code{n} \eqn{\times} \code{n} connectivity matrix.}
}
\value{
Object of class \code{FEBasis}
}
\description{
This function initialises an object of class \code{FEBasis} which defines a set of `tent' basis functions over a pre-specified triangulation in 2-D
}
\examples{
## See vignette
}
\keyword{basis}
\keyword{elements,}
\keyword{finite}
\keyword{functions}

