% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bivariate_utils.R
\name{logdet}
\alias{logdet}
\title{Find the log determinant}
\usage{
logdet(L)
}
\arguments{
\item{L}{the Cholesky factor of Q}
}
\description{
Find the log determinant of a matrix Q from its Cholesky factor L.
}
\examples{
require(Matrix)
Q <- sparseMatrix(i=c(1,1,2,2),j=c(1,2,1,2),x=c(0.1,0.2,0.2,1))
logdet(chol(Q))
}

