% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,FEBasis,character-method}
\title{Plot meshes and observations}
\usage{
plot(x, y, ...)

\S4method{plot}{FEBasis,character}(x, y, ...)
}
\arguments{
\item{x}{a mesh object}

\item{...}{other parameters used to configure the plot. These include \code{max} (upperbound on colour scale) and \code{min} (lowerbound on colour scale)}
}
\value{
a ggplot2 object
}
\description{
This function is the general one used for plotting/visualising objects mesh objects.
}
\examples{
## See vignette min_max_T
}

