% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Finite_elements.R
\docType{methods}
\name{[<--methods}
\alias{[<-,Basis,ANY,ANY,ANY-method}
\alias{[<-,Basis,ANY,ANY-method}
\alias{[<--methods}
\title{Assignment methods}
\usage{
\S4method{[}{Basis,ANY,ANY,ANY}(x, i, j) <- value
}
\description{
Methods for \code{"[<-"}, i.e., extraction or subsetting of elements in the data frame of the mesh object
}
\examples{
## See vignette min_max_T
}

