% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_classes.R
\docType{class}
\name{FEBasis-class}
\alias{FEBasis-class}
\title{@docType class
 @title Finite element basis}
\description{
FEBasis inherits from the virtual class Basis. A finite element basis is initialised using the function \code{initFEbasis}.
}
\keyword{Basis}
\keyword{functions}

